BIG Fix for DB import issues
When the database is upgraded, a blank database is created (after a backup is made of the existing). The code then adds in all the control/settings to this blank database, however it was ignoring any field in setupform/etc that was blank. This caused the key for that control to not be added to the blank database, and prevented the merge from importing the old setting. The merge iterates through the keys in the new db and pulls over any matches from the old db. A massive issue that has probably been around for quite a while.