fix for recovering existing calibration data
Issue caused by the code requiring the same number of calibration entries from the loaded database as there are models. Due to the new ANVELINAPRO3 this code would be prevented, essentially losing the calibration data for the existing models and returning to defaults. Issue discovered by Tony EI7BMB in the Discord testing group.