Removed references to IARU R1 meter (read notes)
Removed references to the IARU R1 recommendations as this is NOT a true implementation of the IARU R1 recommendation for s-meters which are defined as :

 "Standardisation of S-meter readings
• One S-unit corresponds to a signal level difference of 6 dB,
• On the bands below 30 MHz a meter deviation of S-9 corresponds to an available power of -73 dBm from a continuous wave signal generator connected to the receiver input terminals,
• On the bands above 30 MHz this available power shall be -93 dBm,
• The metering system shall be based on quasi-peak detection with an attack time of 10 msec ∀ 2 msec and a decay time constant of at least 500 msec."

If we look at each point in turn :

- S-unit corresponds to 6dB - accepted
- S-9 corresponds to -73dBm for HF - accepted
- S-9 corresponds to -93dBm for VHF/UHF - accepted
- Time constants for the meter - accepted, rise quickly to capture something near the signal peak amplitude, fall slower

It does NOT define bandwidth or how the reading is to be derived, and probably takes for granted the method used at time of writing, ie. a DC detector measuring filtered RF bandwidth in an analogue radio. This dc detector method would be equivalent to integrated (total) power over the bandpass in wdsp existing meters.