symbol update + channelmaster dll version increment
This channelmaster.dll version increment was missed in 2.10.3.9. Consequently there will be a fatal error if the old version 1.0.0.0 is not successfully removed (by using add/remove uninstall) and 2.10.3.9 is installed. This has now been updated to 1.0.1.0.  You can check in the normal install folder of C:\Program Files\OpenHPSDR\Thetis to check that ChannelMaster.dll has been removed after the uninstall.